Adding test coverage and 'no cover' comments
This code was copied from the data team's report and most of it has no tests.
Some tests have been added to make adding new features and refactoring safer,
but for now I'm skipping coverage checks on the parts that haven't been changed
(although this should be rectified next time that code does need to be changed).