Update test assertions
The team id is now being returned in the post message after an update to
slack-bolt, which broke the tests as they were asserting on the whole dict. The
tests have now been changed to assert only on the things we care about.

The update of slack-bolt happened in version
[1.15.0](https://github.com/ebmdatalab/ebmbot/pull/270).