Put team specific params into function
This moves them out of the slack call, to make that shorter.

This creates a wrapper around generate report with the pipeline team specific
options. It will allow us to move the report generation code out, so that all
teams can use it. It also makes it easy to expand with extra team specific
commands in future, such as listing new issues that have been created but that
aren't on the board, and more.