Remove test warnings and add env var
Freezegun is throwing some deprecation warnings that were making our test
outputs noisy. It's safe for us to ignore this for now, while we wait for
a new version of freezegun.

A new env var has been added and needs to be added to the test config, so we
can run the tests locally without needing to pass the env var in too.