Add a command to remove a job by id
If some unexpected error is encountered during the running of
a job, the job may be updated in the database to indicate that
it's running, but never complete or fail. This prevents the job
being retried, and any further attempts to run the job only update
the job scheduled to run after the current one is done. This adds a
command to remove specific jobs by id (identified by calling status
first).