Retry slack notifications in case of unexpected errors
If the slack client errors in some unexpected way, it can
cause a scheduled job to error without returning any success/fail
status, which means the job gets stuck in the "running" state, even
though it's only been reserved and isn't actually running. Now we
retry a max of 3 times and log the error if we can't do the
notification. This could mean that jobs run and never report
back, if the slack client is more than transiently broken, but
then the chances are that simple commands like help are also
broken and at least the output of jobs should be accessible in
the job logs.