Add the docker user's ssh key with ssh-add
This wasn't needed with fabric3, but now we've replaced it with
fab-classic (and paramiko with paramiko-ng), the ssh key in the
mounted ebmbot user's home directory wasn't being recognised and
it hung asking for a password. It needs to be explicitly added with
ssh-add. I'm not sure the Procfile is the place I really want to
do it, but it's the only way I found that worked. The "release"
command is run before containers are deployed, and a post-deploy
command in app.json also doesn't run in the containers afaict.