Pass thread_ts instead of ts to scheduled jobs
Use thread_ts from the original message instead of ts when
scheduling jobs. thread_ts may not be present when a command is
called directly (not from a thread), which is fine, because the
response will be posted into the channel as a new message, which
is what we want. It will now only be posted as a reply if it was
called from within a thread.