Add `--quiet` flag to NCSO alerts command
Something about the way fabric/ebmbot work means that stderr gets
reported to Slack, interspersed with stdout. This flag disables the
stderr logging to stop the command creating a lot of noise in a Slack.

If there's a simple way to get it just report stdout that would also
solve the problem, but I couldn't immediately see one.

See also:
 * https://github.com/ebmdatalab/openprescribing/pull/4772