Ignore techsupport rota rows without sufficient data
We expect the spreadsheet to contain date,primary name,secondary name in
the first 3 columns. We don't want it to break just because the date column has
been populated ahead of names being filled in for the rota, so just
ignore rows if they don't have data in at least those 3 columns.