Rename get_tech_support_channel to get_channel_id
We'll reuse it later for bennett-admins as well as tech support.

Also avoid calling get_channels for every job; we don't expect the
tech support channel ID to change (and if it does, we'd have to update
the environment variable, which would restart the app and update the
channel ID anyway.