Add a MessageChecker to ensure support messages have been responded to
Runs a MessageChecker in a separate process. This fetches all messages
from today which:
- contain the keyword (tech-support or bennett-admins)
- but have not been reacted to
- and are not from the bot itself
- and are not in the support channel
- and are not DMs (those aren't reposted by the support automation
  anyway)

For each matching message, it adds the relevant reaction and reposts the
message to the support channel (except when the keyword is not in the
message text, i.e. it's a forwarded message or copy/pasted message
link).

This runs every 5 seconds to pick up any support messages that
BennettBot has failed to respond to.