Set local WRITEABLE_DIR to a different location
In production, any files that need to be written by the app (e.g.
the db file, techsupport ooo and files that are fetched by a job -
typically the fabfiles) are written to a separate, writeable
directory. The docker user does not have permission to write to
the main app code, and on dokku3 we mount in a writeable volume for
this purpose.

It was confusing to have these files written to the same workpace
folder that contained committed job files, so we now keep them
separate locally too.