Notify slack if there is an active supression for a merged GitHub PR
If a PR that uses the webserver callback (i.e. OpenPrescribing)
successfully schedules a deploy, check if there are active suppressions
so we can notify the channel. If we don't fo this, there's no way to
tell that the deploy has been successfully scheduled to run later,
unless the user knows to check BennettBot's status.