Extend the text matcher for bot commands
Allows us to match commands that follow an @BennettBot mention, even
if they are preceded by other text. It's quite common to copy/paste a
previous bot command, and accidentally paste a bit of preceding text.
We'll now ignore anything that preceded the @BennettBot mention.

We already handled extra whitespace and a following ".". I haven't
attempted to handle additional text that follows a command, because
a) it's probably not encountered that often and
b) it would be complicated to try and match any job command regex
when we don't have a definite fixed list (we sort of do at any one
time, but it can be added to, and we don't test all combination),
and a number of job commands are subsets of others (e.g. `test hello`
and `test hello [name]`) and so would be dependent on being matched
in order.