Search for missed support messages from today and yesterday
Slack only allows us to search from a date, not a datetime (times
can be used too, but they apply to times on every day).

Sometimes, it seems slack's search just doesn't return messages
for a while. We now make it check for the whole of yesterday
AND today.

We also don't need to give it a "before" search term.