Allow support channel settings to be either channel name OR channel ID
Most slack functions will accept either a channel name or a channel
ID. We always use a channel ID for support channels for the few that
require it. For readability, our settings have used channel names, but the
disadvantage is that if a channel name changes, bennettbot will no
longer be able to post to it.

This allows settings to be either channel name or channel ID. If we
decide to change them to channel IDs in prod, this means channel names
can be changed with no interruption to how the support keywords work.