Move support config to a shared module
Both the bot and the message checker in the dispatcher use elements
of the support config, so move it out into a shared module.

I'm not greatly happy about having a config.py that this lives in,
but I don't want to pollute settings.py with functions that build
config, and I don't want dispatcher.py to be dependent on bot.py

Also fixes a bug with the reaction emoji config, and uses
"support_channel" as the key in the support config to be more
explicit about which channel we're referring to in functions that
use this config.