Use the correct python version when running jobs
In local dev environments, we run with `just`, which explicitly
uses the virtual env bin to run all commands, without the
user needing to manually activate it. However, dispatcher jobs are
run in a subprocess and just run with `python ...` and so don't
automatically use the virtualenv python.

In the Dockerfile (i.e. for production), we prepend the virtualenv
bin path to the PATH environment variable, which ensures that the
virtualenv python will be the first python encountered and used.

We now check for an ABSOLUTE_BIN env variable (set in the justfile, so this
will only apply for local envs, not prod) and ensure that it is first
in the PATH when running a job.

Note that we could set PATH itself in the justfile so that it is passed
through in os.environ to the job subprocess, but if we do,
it'll set the user's global PATH, which is not necessarily what they
want.