Weekly standup rota for Team Rex
We have a standup three days a week and have six people on the rota. So, I've
created a very simple algorithm that pairs people up and assigns them a
particular day of the week. Each person in the pair then alternates running
standup depending on the week.

The weekly summary of the rota shows who's running standup for the current
week.