Calculate "at risk" using retention period info.
The existing calculation:

```
is_dormant = codespace.last_used_days_ago >= threshold_in_days
```

... doesn't take into account that the user that owns the codespace may have
selected to manuallyadjust the codespace expiry period (or remove it entirely),
or to change their personal retention setting. A codespace may be classified as
`is_dormant` even if it will never expire, or if it's due to expire sooner than
we expect.

Instead, calculate based on `retention_period_expiry` which is available in the
record.