Revert to using `print` in output.
Since c086fa0d12aa3c879b9493768c8d4b659ee944ed was pushed, the bot has stopped
printing output when the command is run. From looking at the logged output, I
think this is the culprit: it can't interpret the prettified text as it
expects for display in Slack.