Replace httpretty with mocket
httpretty has not been actively maintained for a while (there are a lot
of outstanding PRs and issues, and the last commit on main was 3 years
ago). We're already pinning it because of a bug, and it's now
incompatible with the latest version of urllib3
https://github.com/gabrielfalcao/HTTPretty/issues/484

mocket does all the same things as httpretty and more (it does have a
plugin that aims to be a dropin replacement for httpretty, but it
doesn't cover all the ways that we use it). In any case, replacing our
use of httpretty with mocket is quite straightforward.