`codespaces at risk`: parametrize threshold.
Allow the Slack request to specify the threshold rather than hard-coding it. We
may fine-tune this depending on how much output we get and what proves most
useful to users. This makes it easier to do that experimentation, change Slack
reminders etc.

I'll update the playbook with the current threshold and rationale. That'll be a
separate PR as it's a separate repo (team-manual).

As discussed in Slack:
https://bennettoxford.slack.com/archives/C069SADHP1Q/p1736434022622129?thread_ts=1736427127.238419&cid=C069SADHP1Q