Tweak output format.
It wasn't as clear and useful as it could be. User and repo are most important
for working out if it needs responding to, so put them first. Include
"Deletion: " text before the datetime to make it clear to what it refers.

As discussed in Slack: https://bennettoxford.slack.com/archives/C069SADHP1Q/p1736765655458669?thread_ts=1736427127.238419&cid=C069SADHP1Q