Add `PersonIterator` metaclass to `People`.
This is to allow `for person in People` or similar to work.

Now `People.all` is not required.

Add ordering to the dataclass to allow the iterator to be sorted for a
consistent ordering.

Note that changing `TEAM_REX` affects `DependabotRotaReporter`.  This might
help avoid change happening without this linkage being noticed.