Populate `human_readable` automatically.
`PersonCollection.__init__()` populates `human_readable` based on attribute
value if not set explicitly. This occurs when the class definition is read.
This avoids the need to explicitly specify it for most people, when it's mostly
duplicative of the attribute value, for example MIKE => Mike.

Cases where the attempted conversion doesn't give the result we want can still
be manually specified in the definition of `People`. For example, with "BEN_BC"
it's not easy to determine programatically that all of the second word should
be capitalized, so we specify explicitly as "Ben BC".